+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Only plot lower triangle for `corPlot`
 - make sure date class is POSIXct, POSIXt in `importAURN`
+- fix problem in `importMeta` for AURN - change in database source format
 
 # Changes in openair 1.8-2